:recycle: Rename searchBarV2 component to searchBar
