+++ src/components/searchBar/delegates.js
