+++ src/components/searchBar/delegates.test.js
