+++ src/components/searchBar/searchBar.test.js
